Added fav column saving + restoring
Added the ability to set your favourite columns and then restore them, also restores them on discover page open.
Refactored code that decodes url to be reusable nicely.
Fixed a bug when using the 'focus on traceid' button when you have some text in the query search box.
Added a random message of the day.